Update vector layout attribute and add helper functions
This PR updates the layout attribute to use the more generic
assembly format for printing and parsing. It also restricts
the dimension names to pre-defined enums instead of strings.
Finally, we add a helper function to return the shape of the
SIMT vector in the order of the dims passes in by the helper
function.